[#4]implement delete empty emelement function

+++ libs/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: takira <takira@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/16 10:46:22 by takira            #+#    #+#             */
-/*   Updated: 2023/01/18 16:38:20 by takira           ###   ########.fr       */
+/*   Updated: 2023/01/19 16:49:26 by takira           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -107,7 +107,7 @@ void		ft_lstdelone(t_list *lst, void (*del)(void *));
 void		ft_lstclear(t_list **lst, void (*del)(void *));
 void		ft_lstiter(t_list *lst, void (*f)(void *));
 
-int			ft_lstsize(t_list *lst);
+size_t		ft_lstsize(t_list *lst);
 
 /* math */
 size_t		minsize(size_t a, size_t b);
